[CM] Inconsistent intptr_t definitions
git-svn-id: https://codex.cro.st.com/svnroot/open64/trunk@3040 9f6548e3-d00b-0410-a34f-e49d3b37daf0

+++ osprey/sunos-cygwin/include/host/H_stdint.h
@@ -10,7 +10,7 @@ typedef u_int16_t uint16_t;
 typedef u_int32_t uint32_t;
 typedef u_int64_t uint64_t;
 
-typedef signed long     intptr_t;
-typedef unsigned long   uintptr_t;
+typedef signed int     intptr_t;
+typedef unsigned int   uintptr_t;
 
 #endif /* H_STDINT_H */
